Extract enums, classes and interfaces from IrImageVector.kt
