+++ components/ir/src/commonMain/kotlin/io/github/composegears/valkyrie/ir/IrFill.kt
@@ -1,0 +1,25 @@
+package io.github.composegears.valkyrie.ir
+
+sealed interface IrFill {
+    data class Color(val irColor: IrColor) : IrFill
+
+    data class LinearGradient(
+        val startY: Float,
+        val startX: Float,
+        val endY: Float,
+        val endX: Float,
+        val colorStops: MutableList<ColorStop> = mutableListOf(),
+    ) : IrFill
+
+    data class RadialGradient(
+        val radius: Float,
+        val centerX: Float,
+        val centerY: Float,
+        val colorStops: MutableList<ColorStop> = mutableListOf(),
+    ) : IrFill
+
+    data class ColorStop(
+        val offset: Float,
+        val irColor: IrColor,
+    )
+}
