+++ components/ir/src/commonMain/kotlin/io/github/composegears/valkyrie/ir/IrPathFillType.kt
@@ -1,0 +1,6 @@
+package io.github.composegears.valkyrie.ir
+
+enum class IrPathFillType {
+    EvenOdd,
+    NonZero,
+}
