+++ components/ir/src/commonMain/kotlin/io/github/composegears/valkyrie/ir/IrStroke.kt
@@ -1,0 +1,5 @@
+package io.github.composegears.valkyrie.ir
+
+sealed interface IrStroke {
+    data class Color(val irColor: IrColor) : IrStroke
+}
