+++ components/ir/src/commonMain/kotlin/io/github/composegears/valkyrie/ir/IrStrokeLineCap.kt
@@ -1,0 +1,7 @@
+package io.github.composegears.valkyrie.ir
+
+enum class IrStrokeLineCap {
+    Butt,
+    Round,
+    Square,
+}
