+++ components/ir/src/commonMain/kotlin/io/github/composegears/valkyrie/ir/IrStrokeLineJoin.kt
@@ -1,0 +1,7 @@
+package io.github.composegears.valkyrie.ir
+
+enum class IrStrokeLineJoin {
+    Miter,
+    Round,
+    Bevel,
+}
